Update patch set 1
Patch Set 1:

This is based on rebasing existing unit test cases from ECOMP code base to ONAP, and we had to first coordinate and run the tests locally. This is in addition to configuring tests with Jenkins (with tox.ini). Hence the reason for the jumbo commit of this seed code.

Patch-set: 1