Update patch set 1
Patch Set 1:

We can add the file, but the main concern is the merge process if the file updates in the future from github... May be add the IDE part o the end of the file? Even though many projects are starting from scratch, I do not want to spend time figuring out patterns for gitignore when github did a fine job.

Patch-set: 1