Update patch set 2
Patch Set 2:

Dileep, the MEMORY part is per VNF/VM, so the deletion is correct. I am not sure about flavorLabel. If flavorLabel is a new requirement, it makes sense to add it as a jira (sorry, since we are near code freeze, it makes sense to limit new changes, or at least track them via jiras)

Patch-set: 2