Update patch set 3
Patch Set 3:

As long as test/policy-local-files are correct we should be good.  

test/functest/scripts/start-simulators.sh copies all the policies and puts it in the simulator's policy directory.

Same is the case with the dockerized version of the simulator.  ref -> test/functest/simulators/build_sim_image.sh

Patch-set: 3
Label: Code-Review=0