Update patch set 1
Patch Set 1:

(12 comments)

From the commit statement, it seems that the policy framework requires tosca model file names in a specific format and therefore, a new set of files has been created under a separate folder. This would lead us to have two copies of tosca models. We may need to handle this redundancy and keep only one copy of the models in next release if not possible now.

Patch-set: 1
Label: Verified=0