Update patch set 3
Patch Set 3:

(7 comments)

Avteet, please find my replies. And also, after the policy changes, the new policy type is being sent as the constraint type to conductor. Conductor won't be able to understand it. We should have a mapping of policy type and constraint type in the conductor interface.

Patch-set: 3
Label: Code-Review=0