Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> A few comments/queries:
> 1. "Type" attribute in NxTTerminateRequest indicates if it is NST or NSST.
Okay, I will change it to enum
> 2. I assume SelectNSI will be called for both NSI selection, as well as NSSI selection in case of RAN subnet (for RAN NF + FH + MH). Is this correct?
Yes.For both the cases this API will be called.
> 3. What will the async response for SelectNSSI be, if no solution is found? Should we include an attribute to state "no existing NSSIs found", or solutions will be returned as an empty field?
The solution object will contain an empty list. In wiki, the carnality is documented as 0..N. I will add it in the description.
> 4. domainType - I assume this will be aligned with the model definition, and can be AN, AN-NF, TN-FH, TN-MH and TN-BH, CN (basically 6 values).
Initially, it was defined as enum. Later I changed after krishna suggested it to be model driven.

Patch-set: 3