Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> 1 comment: Can you pl. clarify if the scenario of having multiple links between 2 nodes is covered (I had a look into apps/route/optimizers/inter_domain_route_opt.py, but it seems to be not covered). Pl. refer to the discussion we had w.r.to realizing this, for example, including a "virtual"/"dummy" node for the 2nd link, etc.

Hi Swami,

Based on the discussion we had in the first call, it was assumed that the current minizinc template that's implemented for the route selection cannot handle multiple links between controllers scenario.
I tested the template with sample data wherein I fed the template with different links including multiple links between same controllers, so the output was such that if two links between two controllers have the same specification then minizinc automatically chooses one for us. So right now the filtration of the links is based on the bandwidth availability of the ports only. Have tested this feature with real topology data from AAI as well. Probably I could change the AAI sample data (including multiple links between controllers) in the unit test cases to show this feature.

Patch-set: 3