Update patch set 8
Patch Set 8:

Hi Shankar, Vikas,

I have move the mapping to the config as you suggested. Can you review it once and merge it if okay. Currently, this is the only open review for slicing. If it is merged, we can proceed to create a release image, so that it will have MDONS and slicing changes.

Patch-set: 8
Reviewer: Gerrit User 2671 <2671@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>